requirements.txt
Flask
Flask-Session
gunicorn
python-dotenv
google-generativeai
fitz # This is the PyMuPDF library
PyMuPDF # Sometimes needed explicitly
python-docx